Staging environment

+++ buildSrc/src/main/java/se/nbis/lega/deployment/cluster/ClusterTask.java
@@ -1,14 +1,17 @@
 package se.nbis.lega.deployment.cluster;
 
+import lombok.extern.slf4j.Slf4j;
 import se.nbis.lega.deployment.Groups;
 import se.nbis.lega.deployment.LocalEGATask;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
+@Slf4j
 public abstract class ClusterTask extends LocalEGATask {
 
     public ClusterTask() {
@@ -30,23 +33,41 @@ public abstract class ClusterTask extends LocalEGATask {
     }
 
     protected Map<String, String> createMachineVirtualBox(String name) throws IOException {
+        log.info("createMachineVirtualBox");
         exec(true, "docker-machine create", "--driver", "virtualbox", name);
         return getMachines(name).get(name);
     }
 
-    protected Map<String, String> createMachineOpenStack(String name, String openStackConfig)
-        throws IOException {
+    protected Map<String, String> createMachineOpenStack(String name, String openStackConfig) throws IOException {
         return createMachineOpenStack(name, readFileAsMap(new File(openStackConfig)));
     }
 
     protected Map<String, String> createMachineOpenStack(String name, Map<String, String> openStackEnvironment)
-        throws IOException {
-        exec(true, openStackEnvironment, "docker-machine create", "--driver", "openstack", name);
+                    throws IOException {
+        log.info("createMachineOpenStack");
+        try {
+            List<String> output =
+                            exec(true, openStackEnvironment, "docker-machine create", "--driver", "openstack", name);
+            for (Iterator<String> iterator = output.iterator(); iterator.hasNext();) {
+                String line = iterator.next();
+                log.info(line);
+            }
+            if (output.get(0).startsWith("Error checking TLS connection")) {
+                regenrateCerts(name, openStackEnvironment);
+            }
+        } catch (Exception e) {
+            log.error("Error Creating openStack machine: " + e.getMessage());
+            e.printStackTrace();
+            throw e;
+        }
         return getMachines(name).get(name);
     }
 
     protected void removeMachine(String name) throws IOException {
-        exec("docker-machine rm -y", name);
+        List<String> machines = exec("docker-machine ls --filter name=", name);
+        if (machines.size() > 1) {
+            exec("docker-machine rm -y", name);
+        }
     }
 
 }
